renamed barklog pkg to models, renamed Database struct to BarkPostgresDb
